arbo: fix the clean up of nodes after delete/update/add
Remove all intermediate nodes from the database once
they are not necessary anymore.

Break up the code into smaller Go files.
