+++ tree/arbo/tree.go
@@ -17,8 +17,6 @@ import (
 	"encoding/hex"
 	"fmt"
 	"io"
-	"math"
-	"runtime"
 	"sync"
 
 	"go.vocdoni.io/dvote/db"
@@ -142,8 +140,12 @@ func NewTreeWithTx(wTx db.WriteTx, cfg Config) (*Tree, error) {
 		cfg.ThresholdNLeafs = DefaultThresholdNLeafs
 	}
 
-	t := Tree{db: cfg.Database, maxLevels: cfg.MaxLevels,
-		thresholdNLeafs: cfg.ThresholdNLeafs, hashFunction: cfg.HashFunction}
+	t := Tree{
+		db:              cfg.Database,
+		maxLevels:       cfg.MaxLevels,
+		thresholdNLeafs: cfg.ThresholdNLeafs,
+		hashFunction:    cfg.HashFunction,
+	}
 
 	t.emptyHash = make([]byte, t.hashFunction.Len()) // empty
 	var err error
@@ -207,320 +209,6 @@ type Invalid struct {
 	Error error
 }
 
-// AddBatch adds a batch of key-values to the Tree. Returns an array containing
-// the indexes of the keys failed to add. Supports empty values as input
-// parameters, which is equivalent to 0 valued byte array.
-func (t *Tree) AddBatch(keys, values [][]byte) ([]Invalid, error) {
-	wTx := t.db.WriteTx()
-	defer wTx.Discard()
-
-	invalids, err := t.AddBatchWithTx(wTx, keys, values)
-	if err != nil {
-		return invalids, err
-	}
-	return invalids, wTx.Commit()
-}
-
-// AddBatchWithTx does the same than the AddBatch method, but allowing to pass
-// the db.WriteTx that is used. The db.WriteTx will not be committed inside
-// this method.
-func (t *Tree) AddBatchWithTx(wTx db.WriteTx, keys, values [][]byte) ([]Invalid, error) {
-	t.Lock()
-	defer t.Unlock()
-
-	if !t.editable() {
-		return nil, ErrSnapshotNotEditable
-	}
-
-	e := []byte{}
-	// equal the number of keys & values
-	if len(keys) > len(values) {
-		// add missing values
-		for i := len(values); i < len(keys); i++ {
-			values = append(values, e)
-		}
-	} else if len(keys) < len(values) {
-		// crop extra values
-		values = values[:len(keys)]
-	}
-
-	nLeafs, err := t.GetNLeafsWithTx(wTx)
-	if err != nil {
-		return nil, err
-	}
-	if nLeafs > t.thresholdNLeafs {
-		return t.addBatchInDisk(wTx, keys, values)
-	}
-	return t.addBatchInMemory(wTx, keys, values)
-}
-
-func (t *Tree) addBatchInDisk(wTx db.WriteTx, keys, values [][]byte) ([]Invalid, error) {
-	nCPU := flp2(runtime.NumCPU())
-	if nCPU == 1 || len(keys) < nCPU {
-		var invalids []Invalid
-		for i := 0; i < len(keys); i++ {
-			if err := t.addWithTx(wTx, keys[i], values[i]); err != nil {
-				invalids = append(invalids, Invalid{i, err})
-			}
-		}
-		return invalids, nil
-	}
-
-	kvs, invalids, err := keysValuesToKvs(t.maxLevels, keys, values)
-	if err != nil {
-		return nil, err
-	}
-
-	buckets := splitInBuckets(kvs, nCPU)
-
-	root, err := t.RootWithTx(wTx)
-	if err != nil {
-		return nil, err
-	}
-
-	l := int(math.Log2(float64(nCPU)))
-	subRoots, err := t.getSubRootsAtLevel(wTx, root, l+1)
-	if err != nil {
-		return nil, err
-	}
-	if len(subRoots) != nCPU {
-		// Already populated Tree but Unbalanced.
-
-		// add one key at each bucket, and then continue with the flow
-		for i := 0; i < len(buckets); i++ {
-			// add one leaf of the bucket, if there is an error when
-			// adding the k-v, try to add the next one of the bucket
-			// (until one is added)
-			inserted := -1
-			for j := 0; j < len(buckets[i]); j++ {
-				if newRoot, err := t.add(wTx, root, 0,
-					buckets[i][j].k, buckets[i][j].v); err == nil {
-					inserted = j
-					root = newRoot
-					break
-				}
-			}
-
-			// remove the inserted element from buckets[i]
-			if inserted != -1 {
-				buckets[i] = append(buckets[i][:inserted], buckets[i][inserted+1:]...)
-			}
-		}
-		subRoots, err = t.getSubRootsAtLevel(wTx, root, l+1)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if len(subRoots) != nCPU {
-		return nil, fmt.Errorf("this error should not be reached."+
-			" len(subRoots) != nCPU, len(subRoots)=%d, nCPU=%d."+
-			" Please report it in a new issue:"+
-			" https://go.vocdoni.io/dvote/tree/arbo/issues/new", len(subRoots), nCPU)
-	}
-
-	invalidsInBucket := make([][]Invalid, nCPU)
-	txs := make([]db.WriteTx, nCPU)
-	for i := 0; i < nCPU; i++ {
-		txs[i] = t.db.WriteTx()
-		err := txs[i].Apply(wTx)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	var wg sync.WaitGroup
-	wg.Add(nCPU)
-	for i := 0; i < nCPU; i++ {
-		go func(cpu int) {
-			// use different wTx for each cpu, after once all
-			// are done, iter over the cpuWTxs and copy their
-			// content into the main wTx
-			for j := 0; j < len(buckets[cpu]); j++ {
-				newSubRoot, err := t.add(txs[cpu], subRoots[cpu],
-					l, buckets[cpu][j].k, buckets[cpu][j].v)
-				if err != nil {
-					invalidsInBucket[cpu] = append(invalidsInBucket[cpu],
-						Invalid{buckets[cpu][j].pos, err})
-					continue
-				}
-				// if there has not been errors, set the new subRoots[cpu]
-				subRoots[cpu] = newSubRoot
-			}
-			wg.Done()
-		}(i)
-	}
-	wg.Wait()
-
-	for i := 0; i < nCPU; i++ {
-		if err := wTx.Apply(txs[i]); err != nil {
-			return nil, err
-		}
-		txs[i].Discard()
-	}
-
-	for i := 0; i < len(invalidsInBucket); i++ {
-		invalids = append(invalids, invalidsInBucket[i]...)
-	}
-
-	newRoot, err := t.upFromSubRoots(wTx, subRoots)
-	if err != nil {
-		return nil, err
-	}
-
-	// update dbKeyNLeafs
-	if err := t.SetRootWithTx(wTx, newRoot); err != nil {
-		return nil, err
-	}
-
-	// update nLeafs
-	if err := t.incNLeafs(wTx, len(keys)-len(invalids)); err != nil {
-		return nil, err
-	}
-
-	return invalids, nil
-}
-
-func (t *Tree) upFromSubRoots(wTx db.WriteTx, subRoots [][]byte) ([]byte, error) {
-	// is a method of Tree just to get access to t.hashFunction and
-	// t.emptyHash.
-
-	// go up from subRoots to up, storing nodes in the given WriteTx
-	// once up at the root, store it in the WriteTx using the dbKeyRoot
-	if len(subRoots) == 1 {
-		return subRoots[0], nil
-	}
-	// get the subRoots values to know the node types of each subRoot
-	nodeTypes := make([]byte, len(subRoots))
-	for i := 0; i < len(subRoots); i++ {
-		if bytes.Equal(subRoots[i], t.emptyHash) {
-			nodeTypes[i] = PrefixValueEmpty
-			continue
-		}
-		v, err := wTx.Get(subRoots[i])
-		if err != nil {
-			return nil, err
-		}
-		nodeTypes[i] = v[0]
-	}
-
-	var newSubRoots [][]byte
-	for i := 0; i < len(subRoots); i += 2 {
-		if (bytes.Equal(subRoots[i], t.emptyHash) && bytes.Equal(subRoots[i+1], t.emptyHash)) ||
-			(nodeTypes[i] == PrefixValueLeaf && bytes.Equal(subRoots[i+1], t.emptyHash)) {
-			// when both sub nodes are empty, the parent is also empty
-			// or
-			// when 1st sub node is a leaf but the 2nd is empty, the
-			// leaf is used as 'parent'
-
-			newSubRoots = append(newSubRoots, subRoots[i])
-			continue
-		}
-		if bytes.Equal(subRoots[i], t.emptyHash) && nodeTypes[i+1] == PrefixValueLeaf {
-			// when 2nd sub node is a leaf but the 1st is empty,
-			// the leaf is used as 'parent'
-			newSubRoots = append(newSubRoots, subRoots[i+1])
-			continue
-		}
-
-		k, v, err := t.newIntermediate(subRoots[i], subRoots[i+1])
-		if err != nil {
-			return nil, err
-		}
-		// store k-v to db
-		if err = wTx.Set(k, v); err != nil {
-			return nil, err
-		}
-		newSubRoots = append(newSubRoots, k)
-	}
-
-	return t.upFromSubRoots(wTx, newSubRoots)
-}
-
-func (t *Tree) getSubRootsAtLevel(rTx db.Reader, root []byte, l int) ([][]byte, error) {
-	// go at level l and return each node key, where each node key is the
-	// subRoot of the subTree that starts there
-
-	var subRoots [][]byte
-	err := t.iterWithStop(rTx, root, 0, func(currLvl int, k, v []byte) bool {
-		if currLvl == l && !bytes.Equal(k, t.emptyHash) {
-			subRoots = append(subRoots, k)
-		}
-		if currLvl >= l {
-			return true // to stop the iter from going down
-		}
-		return false
-	})
-
-	return subRoots, err
-}
-
-func (t *Tree) addBatchInMemory(wTx db.WriteTx, keys, values [][]byte) ([]Invalid, error) {
-	vt, err := t.loadVT(wTx)
-	if err != nil {
-		return nil, err
-	}
-
-	invalids, err := vt.addBatch(keys, values)
-	if err != nil {
-		return nil, err
-	}
-
-	// once the VirtualTree is build, compute the hashes
-	pairs, err := vt.computeHashes()
-	if err != nil {
-		// currently invalids in computeHashes are not counted,
-		// but should not be needed, as if there is an error there is
-		// nothing stored in the db and the error is returned
-		return nil, err
-	}
-
-	// store pairs in db
-	for i := 0; i < len(pairs); i++ {
-		if err := wTx.Set(pairs[i][0], pairs[i][1]); err != nil {
-			return nil, err
-		}
-	}
-
-	// store root (from the vt) to db
-	if vt.root != nil {
-		if err := wTx.Set(dbKeyRoot, vt.root.h); err != nil {
-			return nil, err
-		}
-	}
-
-	// update nLeafs
-	if err := t.incNLeafs(wTx, len(keys)-len(invalids)); err != nil {
-		return nil, err
-	}
-
-	return invalids, nil
-}
-
-// loadVT loads a new virtual tree (vt) from the current Tree, which contains
-// the same leafs.
-func (t *Tree) loadVT(rTx db.Reader) (vt, error) {
-	vt := newVT(t.maxLevels, t.hashFunction)
-	vt.params.dbg = t.dbg
-	var callbackErr error
-	err := t.IterateWithStopWithTx(rTx, nil, func(_ int, k, v []byte) bool {
-		if v[0] != PrefixValueLeaf {
-			return false
-		}
-		leafK, leafV := ReadLeafValue(v)
-		if err := vt.add(0, leafK, leafV); err != nil {
-			callbackErr = err
-			return true
-		}
-		return false
-	})
-	if callbackErr != nil {
-		return vt, callbackErr
-	}
-
-	return vt, err
-}
-
 // Add inserts the key-value into the Tree.  If the inputs come from a
 // *big.Int, is expected that are represented by a Little-Endian byte array
 // (for circom compatibility).
@@ -565,45 +253,7 @@ func (t *Tree) addWithTx(wTx db.WriteTx, k, v []byte) error {
 		return err
 	}
 	// update nLeafs
-	if err = t.incNLeafs(wTx, 1); err != nil {
-		return err
-	}
-	return nil
-}
-
-// keyPathFromKey returns the keyPath and checks that the key is not bigger
-// than maximum key length for the tree maxLevels size.
-// This is because if the key bits length is bigger than the maxLevels of the
-// tree, two different keys that their difference is at the end, will collision
-// in the same leaf of the tree (at the max depth).
-func keyPathFromKey(maxLevels int, k []byte) ([]byte, error) {
-	maxKeyLen := int(math.Ceil(float64(maxLevels) / float64(8)))
-	if len(k) > maxKeyLen {
-		return nil, fmt.Errorf("len(k) can not be bigger than ceil(maxLevels/8), where"+
-			" len(k): %d, maxLevels: %d, max key len=ceil(maxLevels/8): %d. Might need"+
-			" a bigger tree depth (maxLevels>=%d) in order to input keys of length %d",
-			len(k), maxLevels, maxKeyLen, len(k)*8, len(k))
-	}
-	keyPath := make([]byte, maxKeyLen)
-	copy(keyPath, k)
-	return keyPath, nil
-}
-
-// checkKeyValueLen checks the key length and value length. This method is used
-// when adding single leafs and also when adding a batch. The limits of lengths
-// used are derived from the encoding of tree dumps: 1 byte to define the
-// length of the keys (2^8-1 bytes length)), and 2 bytes to define the length
-// of the values (2^16-1 bytes length).
-func checkKeyValueLen(k, v []byte) error {
-	if len(k) > maxUint8 {
-		return fmt.Errorf("len(k)=%v, can not be bigger than %v",
-			len(k), maxUint8)
-	}
-	if len(v) > maxUint16 {
-		return fmt.Errorf("len(v)=%v, can not be bigger than %v",
-			len(v), maxUint16)
-	}
-	return nil
+	return t.incNLeafs(wTx, 1)
 }
 
 func (t *Tree) add(wTx db.WriteTx, root []byte, fromLvl int, k, v []byte) ([]byte, error) {
@@ -618,10 +268,9 @@ func (t *Tree) add(wTx db.WriteTx, root []byte, fromLvl int, k, v []byte) ([]byt
 	path := getPath(t.maxLevels, keyPath)
 
 	// go down to the leaf
-	var siblings [][]byte
-	intermediates := make([][]byte, t.maxLevels)
+	var siblings, intermediates [][]byte
 
-	_, _, siblings, intermediates, err = t.down(wTx, k, root, siblings, intermediates, path, fromLvl, false)
+	_, _, siblings, err = t.down(wTx, k, root, siblings, &intermediates, path, fromLvl, false)
 	if err != nil {
 		return nil, err
 	}
@@ -630,6 +279,7 @@ func (t *Tree) add(wTx db.WriteTx, root []byte, fromLvl int, k, v []byte) ([]byt
 	if err != nil {
 		return nil, err
 	}
+
 	if err := wTx.Set(leafKey, leafValue); err != nil {
 		return nil, err
 	}
@@ -639,78 +289,21 @@ func (t *Tree) add(wTx db.WriteTx, root []byte, fromLvl int, k, v []byte) ([]byt
 		// return the leafKey as root
 		return leafKey, nil
 	}
-	root, err = t.up(wTx, intermediates, leafKey, siblings, path, len(siblings)-1, fromLvl)
+	root, err = t.up(wTx, leafKey, siblings, path, len(siblings)-1, fromLvl)
 	if err != nil {
 		return nil, err
+	}
+
+	if err := deleteNodes(wTx, intermediates); err != nil {
+		return root, fmt.Errorf("error deleting orphan intermediate nodes: %v", err)
 	}
 
 	return root, nil
 }
 
-// downVirtually is used when in a leaf already exists, and a new leaf which
-// shares the path until the existing leaf is being added
-func (t *Tree) downVirtually(siblings [][]byte, oldKey, newKey []byte, oldPath,
-	newPath []bool, currLvl int) ([][]byte, error) {
-	var err error
-	if currLvl > t.maxLevels-1 {
-		return nil, ErrMaxVirtualLevel
-	}
-
-	if oldPath[currLvl] == newPath[currLvl] {
-		siblings = append(siblings, t.emptyHash)
-
-		siblings, err = t.downVirtually(siblings, oldKey, newKey, oldPath, newPath, currLvl+1)
-		if err != nil {
-			return nil, err
-		}
-		return siblings, nil
-	}
-	// reached the divergence
-	siblings = append(siblings, oldKey)
-
-	return siblings, nil
-}
-
 func (t *Tree) newLeafValue(k, v []byte) ([]byte, []byte, error) {
 	t.dbg.incHash()
 	return newLeafValue(t.hashFunction, k, v)
-}
-
-// newLeafValue takes a key & value from a leaf, and computes the leaf hash,
-// which is used as the leaf key. And the value is the concatenation of the
-// inputted key & value. The output of this function is used as key-value to
-// store the leaf in the DB.
-// [     1 byte   |     1 byte    | N bytes | M bytes ]
-// [ type of node | length of key |   key   |  value  ]
-func newLeafValue(hashFunc HashFunction, k, v []byte) ([]byte, []byte, error) {
-	if err := checkKeyValueLen(k, v); err != nil {
-		return nil, nil, err
-	}
-	leafKey, err := hashFunc.Hash(k, v, []byte{1})
-	if err != nil {
-		return nil, nil, err
-	}
-	var leafValue []byte
-	leafValue = append(leafValue, byte(PrefixValueLeaf))
-	leafValue = append(leafValue, byte(len(k)))
-	leafValue = append(leafValue, k...)
-	leafValue = append(leafValue, v...)
-	return leafKey, leafValue, nil
-}
-
-// ReadLeafValue reads from a byte array the leaf key & value
-func ReadLeafValue(b []byte) ([]byte, []byte) {
-	if len(b) < PrefixValueLen {
-		return []byte{}, []byte{}
-	}
-
-	kLen := b[1]
-	if len(b) < PrefixValueLen+int(kLen) {
-		return []byte{}, []byte{}
-	}
-	k := b[PrefixValueLen : PrefixValueLen+kLen]
-	v := b[PrefixValueLen+kLen:]
-	return k, v
 }
 
 func (t *Tree) newIntermediate(l, r []byte) ([]byte, []byte, error) {
@@ -752,34 +345,47 @@ func (t *Tree) UpdateWithTx(wTx db.WriteTx, k, v []byte) error {
 		return err
 	}
 
-	var siblings [][]byte
-	intermediates := make([][]byte, t.maxLevels)
-
-	_, valueAtBottom, siblings, intermediates, err := t.down(wTx, k, root, siblings, intermediates, path, 0, true)
+	var siblings, intermediates [][]byte
+	oldLeafKey, valueAtBottom, siblings, err := t.down(wTx, k, root, siblings, &intermediates, path, 0, true)
 	if err != nil {
 		return err
 	}
+
+	// check if the key is actually the same
 	oldKey, _ := ReadLeafValue(valueAtBottom)
 	if !bytes.Equal(oldKey, k) {
 		return ErrKeyNotFound
 	}
 
+	// delete the old leaf key
+	if err := wTx.Delete(oldLeafKey); err != nil {
+		return fmt.Errorf("error deleting old leaf on update: %w", err)
+	}
+
+	// compute the new leaf key and value
 	leafKey, leafValue, err := t.newLeafValue(k, v)
 	if err != nil {
 		return err
 	}
 
+	// add the new leaf key
 	if err := wTx.Set(leafKey, leafValue); err != nil {
 		return err
 	}
 
 	// go up to the root
+
 	if len(siblings) == 0 {
 		return t.setRoot(wTx, leafKey)
 	}
-	root, err = t.up(wTx, intermediates, leafKey, siblings, path, len(siblings)-1, 0)
+
+	root, err = t.up(wTx, leafKey, siblings, path, len(siblings)-1, 0)
 	if err != nil {
 		return err
+	}
+
+	if err := deleteNodes(wTx, intermediates); err != nil {
+		return fmt.Errorf("error deleting orphan intermediate nodes: %v", err)
 	}
 
 	// store root to db
@@ -819,11 +425,10 @@ func (t *Tree) deleteWithTx(wTx db.WriteTx, k []byte) error {
 	}
 
 	// Navigate down the tree to find the key and collect siblings.
-	var siblings [][]byte
-	intermediates := make([][]byte, t.maxLevels)
+	var siblings, intermediates [][]byte
 
 	path := getPath(t.maxLevels, k)
-	leafKey, _, siblings, intermediates, err := t.down(wTx, k, root, siblings, intermediates, path, 0, true)
+	leafKey, _, siblings, err := t.down(wTx, k, root, siblings, &intermediates, path, 0, true)
 	if err != nil {
 		if err == ErrKeyNotFound {
 			// Key not found, nothing to delete.
@@ -832,9 +437,16 @@ func (t *Tree) deleteWithTx(wTx db.WriteTx, k []byte) error {
 		return err
 	}
 
-	// delete the leaf key
-	if err := wTx.Delete(leafKey); err != nil {
-		return fmt.Errorf("error deleting key %x: %w", leafKey, err)
+	// if the neighbor is not empty, set the leaf key to the empty hash
+	if !bytes.Equal(siblings[len(siblings)-1], t.emptyHash) {
+		if err := wTx.Set(leafKey, t.emptyHash); err != nil {
+			return fmt.Errorf("error setting leaf key %x to empty hash: %w", leafKey, err)
+		}
+	} else {
+		// else delete the leaf key
+		if err := wTx.Delete(leafKey); err != nil {
+			return fmt.Errorf("error deleting leaf key %x: %w", leafKey, err)
+		}
 	}
 
 	// Navigate back up the tree, updating the intermediate nodes.
@@ -842,7 +454,7 @@ func (t *Tree) deleteWithTx(wTx db.WriteTx, k []byte) error {
 		// The tree is empty after deletion.
 		return t.setRoot(wTx, t.emptyHash)
 	}
-	newRoot, err := t.up(wTx, intermediates, t.emptyHash, siblings, path, len(siblings)-1, 0)
+	newRoot, err := t.up(wTx, t.emptyHash, siblings, path, len(siblings)-1, 0)
 	if err != nil {
 		return err
 	}
@@ -852,165 +464,12 @@ func (t *Tree) deleteWithTx(wTx db.WriteTx, k []byte) error {
 		return err
 	}
 
+	if err := deleteNodes(wTx, intermediates); err != nil {
+		return fmt.Errorf("error deleting orphan intermediate nodes: %v", err)
+	}
+
 	// Update the number of leaves.
 	return t.decNLeafs(wTx, 1)
-}
-
-// decNLeafs decreases the number of leaves in the tree.
-func (t *Tree) decNLeafs(wTx db.WriteTx, n int) error {
-	nLeafs, err := t.GetNLeafsWithTx(wTx)
-	if err != nil {
-		return err
-	}
-	return t.setNLeafs(wTx, nLeafs-n)
-}
-
-// GenProof generates a MerkleTree proof for the given key. The leaf value is
-// returned, together with the packed siblings of the proof, and a boolean
-// parameter that indicates if the proof is of existence (true) or not (false).
-func (t *Tree) GenProof(k []byte) ([]byte, []byte, []byte, bool, error) {
-	return t.GenProofWithTx(t.db, k)
-}
-
-// GenProofWithTx does the same than the GenProof method, but allowing to pass
-// the db.ReadTx that is used.
-func (t *Tree) GenProofWithTx(rTx db.Reader, k []byte) ([]byte, []byte, []byte, bool, error) {
-	keyPath, err := keyPathFromKey(t.maxLevels, k)
-	if err != nil {
-		return nil, nil, nil, false, err
-	}
-	path := getPath(t.maxLevels, keyPath)
-
-	root, err := t.RootWithTx(rTx)
-	if err != nil {
-		return nil, nil, nil, false, err
-	}
-
-	// go down to the leaf
-	var siblings [][]byte
-	intermediates := make([][]byte, t.maxLevels)
-
-	_, value, siblings, intermediates, err := t.down(rTx, k, root, siblings, intermediates, path, 0, true)
-	if err != nil {
-		return nil, nil, nil, false, err
-	}
-
-	s, err := PackSiblings(t.hashFunction, siblings)
-	if err != nil {
-		return nil, nil, nil, false, err
-	}
-
-	leafK, leafV := ReadLeafValue(value)
-	if !bytes.Equal(k, leafK) {
-		// key not in tree, proof of non-existence
-		return leafK, leafV, s, false, nil
-	}
-
-	return leafK, leafV, s, true, nil
-}
-
-// PackSiblings packs the siblings into a byte array.
-// [    2 byte   |     2 byte        | L bytes |      S * N bytes    ]
-// [ full length | bitmap length (L) |  bitmap | N non-zero siblings ]
-// Where the bitmap indicates if the sibling is 0 or a value from the siblings
-// array. And S is the size of the output of the hash function used for the
-// Tree. The 2 2-byte that define the full length and bitmap length, are
-// encoded in little-endian.
-func PackSiblings(hashFunc HashFunction, siblings [][]byte) ([]byte, error) {
-	var b []byte
-	var bitmap []bool
-	emptySibling := make([]byte, hashFunc.Len())
-	for i := 0; i < len(siblings); i++ {
-		if bytes.Equal(siblings[i], emptySibling) {
-			bitmap = append(bitmap, false)
-		} else {
-			bitmap = append(bitmap, true)
-			b = append(b, siblings[i]...)
-		}
-	}
-
-	bitmapBytes := bitmapToBytes(bitmap)
-	l := len(bitmapBytes)
-	if l > maxUint16 {
-		return nil, fmt.Errorf("PackSiblings: bitmapBytes length > %v", maxUint16)
-	}
-
-	fullLen := 4 + l + len(b)
-	if fullLen > maxUint16 {
-		return nil, fmt.Errorf("PackSiblings: fullLen > %v", maxUint16)
-	}
-	res := make([]byte, fullLen)
-	binary.LittleEndian.PutUint16(res[0:2], uint16(fullLen)) // set full length
-	binary.LittleEndian.PutUint16(res[2:4], uint16(l))       // set the bitmapBytes length
-	copy(res[4:4+l], bitmapBytes)
-	copy(res[4+l:], b)
-	return res, nil
-}
-
-// UnpackSiblings unpacks the siblings from a byte array.
-func UnpackSiblings(hashFunc HashFunction, b []byte) ([][]byte, error) {
-	// to prevent runtime slice out of bounds error check if the length of the
-	// rest of the slice is at least equal to the encoded full length value
-	if len(b) < 4 {
-		return nil, fmt.Errorf("no packed siblings provided")
-	}
-
-	fullLen := binary.LittleEndian.Uint16(b[0:2])
-	if len(b) != int(fullLen) {
-		return nil, fmt.Errorf("expected len: %d, current len: %d", fullLen, len(b))
-	}
-
-	l := binary.LittleEndian.Uint16(b[2:4]) // bitmap bytes length
-	// to prevent runtime slice out of bounds error check if the length of the
-	// rest of the slice is at least equal to the encoded bitmap length value
-	if len(b) < int(4+l) {
-		return nil, fmt.Errorf("expected len: %d, current len: %d", 4+l, len(b))
-	}
-	bitmapBytes := b[4 : 4+l]
-	bitmap := bytesToBitmap(bitmapBytes)
-	siblingsBytes := b[4+l:]
-	// to prevent a runtime slice out of bounds error, check if the length of
-	// the siblings slice is a multiple of hashFunc length, because the
-	// following loop will iterate over it in steps of that length.
-	if len(siblingsBytes)%hashFunc.Len() != 0 {
-		return nil, fmt.Errorf("bad formated siblings")
-	}
-	iSibl := 0
-	emptySibl := make([]byte, hashFunc.Len())
-	var siblings [][]byte
-	for i := 0; i < len(bitmap); i++ {
-		if iSibl >= len(siblingsBytes) {
-			break
-		}
-		if bitmap[i] {
-			siblings = append(siblings, siblingsBytes[iSibl:iSibl+hashFunc.Len()])
-			iSibl += hashFunc.Len()
-		} else {
-			siblings = append(siblings, emptySibl)
-		}
-	}
-	return siblings, nil
-}
-
-func bitmapToBytes(bitmap []bool) []byte {
-	bitmapBytesLen := int(math.Ceil(float64(len(bitmap)) / 8))
-	b := make([]byte, bitmapBytesLen)
-	for i := 0; i < len(bitmap); i++ {
-		if bitmap[i] {
-			b[i/8] |= 1 << (i % 8)
-		}
-	}
-	return b
-}
-
-func bytesToBitmap(b []byte) []bool {
-	var bitmap []bool
-	for i := 0; i < len(b); i++ {
-		for j := 0; j < 8; j++ {
-			bitmap = append(bitmap, b[i]&(1<<j) > 0)
-		}
-	}
-	return bitmap
 }
 
 // Get returns the value in the Tree for a given key. If the key is not found,
@@ -1038,10 +497,9 @@ func (t *Tree) GetWithTx(rTx db.Reader, k []byte) ([]byte, []byte, error) {
 	}
 
 	// go down to the leaf
-	var siblings [][]byte
-	intermediates := make([][]byte, t.maxLevels)
+	var siblings, intermediates [][]byte
 
-	_, value, _, intermediates, err := t.down(rTx, k, root, siblings, intermediates, path, 0, true)
+	_, value, _, err := t.down(rTx, k, root, siblings, &intermediates, path, 0, true)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -1053,40 +511,13 @@ func (t *Tree) GetWithTx(rTx db.Reader, k []byte) ([]byte, []byte, error) {
 	return leafK, leafV, nil
 }
 
-// CheckProof verifies the given proof. The proof verification depends on the
-// HashFunction passed as parameter.
-func CheckProof(hashFunc HashFunction, k, v, root, packedSiblings []byte) (bool, error) {
-	siblings, err := UnpackSiblings(hashFunc, packedSiblings)
+// decNLeafs decreases the number of leaves in the tree.
+func (t *Tree) decNLeafs(wTx db.WriteTx, n int) error {
+	nLeafs, err := t.GetNLeafsWithTx(wTx)
 	if err != nil {
-		return false, err
+		return err
 	}
-
-	keyPath := make([]byte, int(math.Ceil(float64(len(siblings))/float64(8))))
-	copy(keyPath, k)
-
-	key, _, err := newLeafValue(hashFunc, k, v)
-	if err != nil {
-		return false, err
-	}
-
-	path := getPath(len(siblings), keyPath)
-	for i := len(siblings) - 1; i >= 0; i-- {
-		if path[i] {
-			key, _, err = newIntermediate(hashFunc, siblings[i], key)
-			if err != nil {
-				return false, err
-			}
-		} else {
-			key, _, err = newIntermediate(hashFunc, key, siblings[i])
-			if err != nil {
-				return false, err
-			}
-		}
-	}
-	if bytes.Equal(key, root) {
-		return true, nil
-	}
-	return false, nil
+	return t.setNLeafs(wTx, nLeafs-n)
 }
 
 func (t *Tree) incNLeafs(wTx db.WriteTx, nLeafs int) error {
@@ -1101,10 +532,7 @@ func (t *Tree) incNLeafs(wTx db.WriteTx, nLeafs int) error {
 func (*Tree) setNLeafs(wTx db.WriteTx, nLeafs int) error {
 	b := make([]byte, 8)
 	binary.LittleEndian.PutUint64(b, uint64(nLeafs))
-	if err := wTx.Set(dbKeyNLeafs, b); err != nil {
-		return err
-	}
-	return nil
+	return wTx.Set(dbKeyNLeafs, b)
 }
 
 // GetNLeafs returns the number of Leafs of the Tree.
